Build the postfix expression with a space in between
Extend the QTNode to store an additional piece of data for postfix
expression

+++ SPA/AST.cpp
@@ -79,7 +79,7 @@ bool AST::matchPattern(string nodeP,string queryP, int type ){
    vector< string > queryPattern( (istream_iterator<string>( stream2 )),istream_iterator<string>() ); 
    // type 1  (_"var"_)
    if(type == 1){
-	   for(int i =0;i<nodePattern.size();i++){
+	   for(size_t i =0;i<nodePattern.size();i++){
 		   if(nodePattern[i] ==queryPattern[0]){
 			   return true;
 		   }
@@ -87,7 +87,7 @@ bool AST::matchPattern(string nodeP,string queryP, int type ){
 	   return false;
    }// type 2 (_"var1 + var2"_)
    else{
-	   for(int k=0;k<nodePattern.size();k++){
+	   for(size_t k=0;k<nodePattern.size();k++){
 		   if(nodePattern[k] == queryPattern[0]){
 			   //base case
 			   if((k+2)>nodePattern.size()){
